move watcher from pkg/gateway to pkg/management
