+++ pkg/management/watcher_test.go
@@ -1,4 +1,4 @@
-package gateway
+package management
 
 import (
 	"context"
