+++ plugins/alerting/pkg/alerting/plugin.go
@@ -2,7 +2,7 @@ package alerting
 
 import (
 	"context"
-	"github.com/rancher/opni/pkg/gateway"
+	"github.com/rancher/opni/pkg/management"
 	"sync"
 	"sync/atomic"
 
@@ -56,7 +56,7 @@ type Plugin struct {
 	cortexOpsClient future.Future[cortexops.CortexOpsClient]
 	natsConn        future.Future[*nats.Conn]
 	js              future.Future[nats.JetStreamContext]
-	globalWatchers  gateway.ConditionWatcher
+	globalWatchers  management.ConditionWatcher
 }
 
 func NewPlugin(ctx context.Context) *Plugin {
